+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
-**./src/common/**
-* Commented all script files.
-* Removed 'capitalLetters' RegExp from 'value-prep.js'
+**./src/common/validation-tasks.js**
+* New function: 'writeStringTooLongErrorText'
+	* Writes message for strings that are too long.
+	* Can be added onto existing error messages, such as line streams.
+	* "Example cannot be longer than x characters."
+* Revised functions to use 'writeStringTooLongErrorText'
+	* checkDataLineLengthNumber
+	* checkDataEntryLengthNumber
+	* checkNameLengthNumber